+++ spec/spec_helper.rb
@@ -4,6 +4,8 @@ require 'bundler/setup'
 require 'service_actor'
 require 'pry'
 
+Dir['./spec/examples/*'].sort.each { |f| require f }
+
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure
   config.example_status_persistence_file_path = '.rspec_status'