+++ CHANGELOG.md
@@ -5,6 +5,10 @@ project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
 
+Breaking changes:
+- Move all code inside `ServiceActor`, only exposing a base `Actor`, enabling
+  you to change the default class name.
+
 Added:
 - Rename `Actor::Context` to `Actor::Result`.
 - Deprecate `context.` in favor of `result.`.